feat: add photos, titles, and names on about us (#406)
This change adds the board members, developers and links to their
linkedIn profiles for the ones that are currently available. There are
tests for the additions as well.

Changed the styling for the map container and main app in order to not
have the footer cut through parts of the page.