feat: update user experience on submission form
This change moves the refs outside of the store with a component  ref that uses a lifecycle hook to update the form when everything is ready to avoid a flicker of the form for the user. There is updated user error messaging from the server response with toasts and uses our updated retry method.