fix: improve submission form ui and ux (#887)
* feat: update ServerResponse to match graphQL type sent from backend

Co-authored-by: gminetoma <gminetoma@gmail.com>

* feat: add error messaging for error code from response

* feat: update user experience on submission form

This change moves the refs outside of the store with a component  ref that uses a lifecycle hook to update the form when everything is ready to avoid a flicker of the form for the user. There is updated user error messaging from the server response with toasts and uses our updated retry method.

* feat: add button to submit another doctor

On the thank you screen a user had to go back to home and then click Add
Doctor in the navbar just to submit another doctor. Some users might
know many doctors that they want to submit so that behavior wasn't good
for the user experience.

* chore: update locale files with submit keys

* test: update test for conditional rendering

* feat: implement server error messaging in graphql request

* fix: change edit submission to use updated call

This change was made in this PR due to otherwise the tests and call failing since
it needed more context and to be updated with the new server response
and typings.

* feat: add toast message for successful rejection

* chore: run script to add rejection message keys

---------

Co-authored-by: gminetoma <gminetoma@gmail.com>