feat: implement frontend filter as fake fuzzy search
This was changed since on input it was calling out backend. This would cost a lot of money as we call the database every time and that is how payment works on this project. Along with this we do not have many doctors as of now so this isn't a heavy operation for the frontend