feat: add MSW to handle requests during tests
- Added MSW
- Updated Cypress

Mock Service Worker (MSW) is an API mocking library for browsers or
Node.js.
With MSW, it's possible to intercept any REST, GraphQL, or WebSocket
requests.
It can be integrated via Node.js with Vitest or via a Service Worker
with Cypress.
Documentation: https://mswjs.io/docs/

The mockServiceWorker.js script should be excluded from production.

The current MSW configuration is set to block any requests to the
"find a doc" API while allowing requests to the Auth0 domain.