feat: implement update existing facility functionality (#902)
* feat: add update button to topbar

* feat: add text for language to update button

* chore: add missing keys for update button on topbar

* feat: change updateFacility function to use updated api call

* feat: change update to fit backend with relationship ref array

* feat: implement frontend filter as fake fuzzy search

This was changed since on input it was calling out backend. This would cost a lot of money as we call the database every time and that is how payment works on this project. Along with this we do not have many doctors as of now so this isn't a heavy operation for the frontend

* feat: add key for update and exit in topbar

* chore: run script to add update and exit keys to locales

* feat: add functionality with checks for updating facility

* chore: update color of trash can to red

This was updated since it is used as a button for a destructive action and in UI design red is for destructive actions

* feat: implement view of existing healthcare relationships

* feat: implement healthcare professional card

This was componentized as it can be used in the edit submission form and edit facility section. It will help keep the code a bit shorter while having consistent styling in different parts of the moderation dashboard

* feat: add key for added healthcare professional

* feat: implement relations for removal of HP from facility

* chore: run script to add keys to tagalog

* feat: add watch to remove healthcare professional in ui

* feat: add gray to our theme colors

For the background color of gray inverted was too dark. This failed all the accessibility scores that I check on Figma with the contrast plugin. This plugin checks the scores for the Web Accessibility Initiative Guidelines. This introduces a gray that passes

* style: update styling of healthcare professional card

* refactor: change fake fuzzy search for better return

This refactors the fuzzy search to search all the locale names and have only ids that match when the starts are the same. It also doesn't display healthcare professionals that are part of the facility already to avoid confusion for moderators

* feat: implement logic to display HP on chosen moderators language

* fix: update search to add and subtract healthcare professionals correctly

* feat: add console error messaging to util function