refactor: make a generic component
`ModSearchBar` is now a generic component with a search input and a results list.

Type inference is implemented using 'Generics'. Learn more here:
https://vuejs.org/api/sfc-script-setup.html#generics

Changed i18n keys to `modHealthcareProfessionalSection`.

Removed old i18n keys.

Added a new prop called `fields-to-display-callback`. This callback
receives the item to be displayed and returns an array of strings,
allowing us to display the desired properties to the user.

Fixed a bug where 'currentFacilities' was empty after a page reload.

Moved 'handleSearchInputChange' to 'ModEditHealthcareProfessionalSection'.