feat: implement logic to add locale for name of HP
This change updates the card to display per name and take in another prop as well as adds a mutation observor instead of onUpdate. onUpdate causes event bubbling that can affect the functionality of the multiselect. The watch will give the multiselect functionality once the select elements exist in the dom per the mutation observer.