fix: remove skip from now working test
This test previously was not working due to the button being clicked before the data was loaded. With a fix earlier for this tcondition where Ithe data was being loaded after ata loaded. With the use of lifecycle hooks and displaying after the data is loaded, this behavior no longer happens and this test passes.

Co-authored-by: gminetoma <gminetoma@gmail.com>