feat: add lazy load to i18n
We no longer need the file './i18n.options.ts' to configure i18n.

The configuration is now handled directly in the Nuxt config file.

We are loading the last language used via cookies; therefore, we need to
disable the language selector and display a loading indicator for the user
until the loading process is complete.

The browser's language code uses "-", so we need to apply a small
adjustment to align with our backend typing, which uses "_".

Co-authored-by: Will <williamedwardbrammer@gmail.com>