fix: fix breaking bugs introduced by previous PRs refactored code
There were bugs introduced into the submission form as we were updatin
g api calls and also with the resetting of the modal refs. This adds re
setting and error handling for toasts within the component to keep logi
c in the same component. It also resets the modal refs on any action wi
th the modal to prevent inconsistencies