feat: add click functionality to autofill button
Added a handler function which runs when autofill button is clicked. The function makes a call to moderationSubmissionStores update submission action. Currently this is proving difficult to test as we don't have a google maps api key. Right now the request returns a 402 error. Currently we don't have access to autofillPlaceFromSubmissionUrl on the frontend so the button will work even if the submission has already autofilled from places, also currently the google map url is hardcoded as the dummy submissions don't have valid map urls.

Co-authored-by: Matthew Kurtti <matthew.b.kurtti@gmail.com>